backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.4,<5.5
django-cors-headers>=4.7,<4.8
mysqlclient>=2.2,<2.3
cos-python-sdk-v5>=1.9,<1.10
